backend/requirements.txt
Flask
pytesseract
Pillow
PyJWT
pdf2image
Flask-Limiter
duckduckgo-search
transformers
torch
google-auth
